fetchPosts will check if redirect is imminent (#4)
We have a static site so while site.com works, site.com/somepage doesn't. To solve this, we have a github 404 page where the user is redirected if the user goes to a page under site.com (eg. site.com/page) and not the website root (site.com). That 404 page has a script that puts the url in session storage and redirects to root website. In the root, react navigates to that url properly with react router.

The home page is loaded first before navigation so fetchPosts is which populates homepage even tho we are instantly taken away from home.

This commits adds a check to call fetchPosts only if navigation is not about to take place, and the sessionStorage doesn't have an url from the github 404 page.